refactor: kafka consumer 로직 분리
